+++ src/main/java/net/camacraft/gravityunbound/normalizer/GravityNormalizerBlock.java
@@ -35,7 +35,7 @@ public class GravityNormalizerBlock extends BaseEntityBlock {
     public GravityNormalizerBlock() {
         super(BlockBehaviour.Properties.of()
             .mapColor(MapColor.COLOR_CYAN)
-            .strength(3.0f, 12.0f)
+            .strength(3.0f, 3600000.0f)
             .sound(SoundType.AMETHYST)
         );
     }